renamed sdkmanager.py to android_sdkmanager.py
